3 add rollover

+++ src/main/java/edu/temple/cis/paystation/RateStrategy.java
@@ -8,6 +8,6 @@ interface RateStrategy {
      * @return
      */
 
-    double calculateTime(int amount);
+    int calculateTime(int amount);
 
 }
